refactor: separate logic from style
